Model (リファクタ)：Neko モデルを NekoGacha モデルから分離
